+++ lisp/cider-apropos.el
@@ -143,7 +143,7 @@ optionally search doc strings (based on DOCS-P), include private vars
                  (y-or-n-p "Include private symbols? ")
                  (y-or-n-p "Case-sensitive? ")))))
   (cider-ensure-connected)
-  (cider-ensure-op-supported "apropos")
+  (cider-ensure-op-supported "cider/apropos")
   (if-let* ((summary (cider-apropos-summary
                       query ns docs-p privates-p case-sensitive-p))
             (results (cider-sync-request:apropos query ns docs-p privates-p case-sensitive-p)))
@@ -155,7 +155,7 @@ optionally search doc strings (based on DOCS-P), include private vars
   "Shortcut for (cider-apropos <query> nil t)."
   (interactive)
   (cider-ensure-connected)
-  (cider-ensure-op-supported "apropos")
+  (cider-ensure-op-supported "cider/apropos")
   (cider-apropos (read-string "Search for Clojure documentation (a regular expression): ") nil t))
 
 (defun cider-apropos-act-on-symbol (symbol)
@@ -189,7 +189,7 @@ optionally search doc strings (based on DOCS-P), include private vars
                  (y-or-n-p "Include private symbols? ")
                  (y-or-n-p "Case-sensitive? ")))))
   (cider-ensure-connected)
-  (cider-ensure-op-supported "apropos")
+  (cider-ensure-op-supported "cider/apropos")
   (if-let* ((summary (cider-apropos-summary
                       query ns docs-p privates-p case-sensitive-p))
             (results (mapcar (lambda (r) (nrepl-dict-get r "name"))
@@ -202,7 +202,7 @@ optionally search doc strings (based on DOCS-P), include private vars
   "Shortcut for (cider-apropos-select <query> nil t)."
   (interactive)
   (cider-ensure-connected)
-  (cider-ensure-op-supported "apropos")
+  (cider-ensure-op-supported "cider/apropos")
   (cider-apropos-select (read-string "Search for Clojure documentation (a regular expression): ") nil t))
 
 (provide 'cider-apropos)